Split config.py into config_utils.py
