+++ manim/logger.py
@@ -1,31 +1,43 @@
-import logging
 import configparser
+import logging
 
-from rich.logging import RichHandler
 from rich.console import Console
+from rich.logging import RichHandler
 from rich.theme import Theme
-#from .config import config
+from .config_utils import config_parser,successfully_read_files
 
-#print(config)
 def parseTheme(fp):
-    config = configparser.ConfigParser()
-    config.read(fp)
-    theme=dict(config['log.color'])
-    customTheme=Theme(theme)
+    config_parser.read(fp)
+    theme = dict(config_parser["log.color"])
+    customTheme = Theme(theme)
     return customTheme
+
 try:
-    customTheme =  parseTheme("manim.cfg")
+    customTheme = parseTheme(successfully_read_files)
     console = Console(theme=customTheme)
 except KeyError:
     console = Console()
 
-RichHandler.KEYWORDS = ['Played', 'animations', 'scene', 'Reading', 'Writing', 'script', 'arguments', 'Invalid', 'Aborting', 'module', 'File', 'Rendering']
+#These keywords Are Highlighted specially.
+RichHandler.KEYWORDS = [
+    "Played",
+    "animations",
+    "scene",
+    "Reading",
+    "Writing",
+    "script",
+    "arguments",
+    "Invalid",
+    "Aborting",
+    "module",
+    "File",
+    "Rendering",
+]
 logging.basicConfig(
     level="NOTSET",
     format="%(message)s",
     datefmt="[%X]",
-    handlers=[RichHandler(console=console)]
+    handlers=[RichHandler(console=console)],
 )
 
 logger = logging.getLogger("rich")
-
